add funcionalidades de JOptionPane confirm Dialog

+++ src/View/TelaCadUsuario.java
@@ -523,12 +523,16 @@ public class TelaCadUsuario extends javax.swing.JFrame {
     }//GEN-LAST:event_btnSalvarActionPerformed
 
     private void btnExcluirActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btnExcluirActionPerformed
-        Usuario obj = new Usuario();
-        obj.setId(Integer.parseInt(txtCodigo.getText()));
-        UsuarioDAO dao = new UsuarioDAO();
-        dao.excluirUsuario(obj);
-        new Ferramentas().LimpaCampo(painelDados);
-        desabilitaBotoes();
+        int op = JOptionPane.showConfirmDialog(null, "Deseja Sair?", "Confirmar ação", JOptionPane.YES_NO_OPTION, JOptionPane.QUESTION_MESSAGE);
+        if (op == JOptionPane.YES_OPTION) {
+            Usuario obj = new Usuario();
+            obj.setId(Integer.parseInt(txtCodigo.getText()));
+            UsuarioDAO dao = new UsuarioDAO();
+            dao.excluirUsuario(obj);
+            new Ferramentas().LimpaCampo(painelDados);
+            desabilitaBotoes();
+        }
+
 
     }//GEN-LAST:event_btnExcluirActionPerformed
 
@@ -561,7 +565,10 @@ public class TelaCadUsuario extends javax.swing.JFrame {
     }//GEN-LAST:event_btnAlterarActionPerformed
 
     private void btnSairActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btnSairActionPerformed
-        dispose();
+        int op = JOptionPane.showConfirmDialog(null, "Deseja Sair?", "Confirmar ação", JOptionPane.YES_NO_OPTION, JOptionPane.QUESTION_MESSAGE);
+        if (op == JOptionPane.YES_OPTION) {
+            dispose();
+        }
     }//GEN-LAST:event_btnSairActionPerformed
 
     private void formWindowActivated(java.awt.event.WindowEvent evt) {//GEN-FIRST:event_formWindowActivated
